refactor: change some imports, add todos
